fix: Corregir parpadeo en home y mejorar navegación en mensajes de soporte
**Problemas corregidos:**
1. Doble loader en home - Movido showLoader=false al final de loadResources()
2. Navegación de mensajes de soporte - Ahora usa Navigator.push en lugar de cambio de estado

**Cambios realizados:**
- driver_home.dart: Mover showLoader=false al bloque finally para evitar parpadeo
- Crear nueva página: create_support_message_page.dart para formulario de nuevo mensaje
- support_messages_unified_page.dart:
  - Remover todo el código del formulario (ahora está en página separada)
  - Agregar _buildEmptyState() para cuando no hay mensajes
  - _navigateToCreateMessage() usa Navigator.push para navegación real
  - Botón físico de atrás ahora funciona correctamente
  - Simplificar estado (remover isCreatingMessage, categories, message, etc.)
