+++ lib/Pages/support_messages_unified_page.dart
@@ -1,5 +1,6 @@
 import 'package:eta_school_app/Models/HelpMessageModel.dart';
 import 'package:eta_school_app/Pages/help_message_page.dart';
+import 'package:eta_school_app/Pages/create_support_message_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:eta_school_app/API/client.dart';
@@ -24,21 +25,12 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
 
   // Estados de la página
   bool showLoader = true;
-  bool isCreatingMessage = false;
 
   // Lista de mensajes
   List<HelpMessageModel>? messagesList = [];
   String? filterStatus;
 
-  // Formulario
-  String message = '';
-  List<SupportHelpCategory>? categories;
-  int categoryId = 0;
-  List<String> priorities = <String>['normal', 'high', 'low'];
-  String priority = 'normal';
-
   // Controladores
-  final TextEditingController _messageController = TextEditingController();
   late AnimationController _fabAnimationController;
   late Animation<double> _fabAnimation;
 
@@ -59,7 +51,6 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
   @override
   void dispose() {
     _fabAnimationController.dispose();
-    _messageController.dispose();
     super.dispose();
   }
 
@@ -68,17 +59,12 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
       showLoader = true;
     });
 
-    await Future.wait([
-      loadMessages(),
-      loadCategories(),
-    ]);
+    await loadMessages();
 
     setState(() {
       showLoader = false;
       // Mostrar FAB después de cargar
-      if (messagesList!.isNotEmpty) {
-        _fabAnimationController.forward();
-      }
+      _fabAnimationController.forward();
     });
   }
 
@@ -93,88 +79,21 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
     }
   }
 
-  Future<void> loadCategories() async {
-    try {
-      final result = await httpService.supportHelpCategory();
-      setState(() {
-        categories = result;
-        if (result.isNotEmpty) {
-          categoryId = result[0].id!;
-        }
-      });
-    } catch (e) {
-      print('[SupportMessages.loadCategories.error] $e');
-    }
-  }
-
   Future<void> _refreshData() async {
     await loadInitialData();
   }
 
-  void _toggleCreateMessage() {
-    setState(() {
-      isCreatingMessage = !isCreatingMessage;
-      if (!isCreatingMessage) {
-        _messageController.clear();
-        message = '';
-        priority = 'normal';
-        if (categories != null && categories!.isNotEmpty) {
-          categoryId = categories![0].id!;
-        }
-      }
-    });
-  }
-
-  Future<void> sendMessage() async {
-    if (message.trim().isEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text(lang.translate('Please enter a message')),
-          backgroundColor: Colors.orange,
-        ),
-      );
-      return;
-    }
-
-    setState(() {
-      showLoader = true;
-    });
-
-    try {
-      final newHelpMessageModel = await httpService.sendMessage(
-        categoryId,
-        message,
-        priorities.indexOf(priority)
-      );
-
-      // Navegar a la página del mensaje
-      Get.to(() => HelpMessagePage(message: newHelpMessageModel));
-
-      // Recargar mensajes
-      await loadMessages();
-
-      setState(() {
-        showLoader = false;
-        isCreatingMessage = false;
-        _messageController.clear();
-        message = '';
-        // Mostrar FAB si ahora hay mensajes
-        if (messagesList!.isNotEmpty) {
-          _fabAnimationController.forward();
-        }
-      });
-    } catch (e) {
-      print('[SupportMessages.sendMessage.error] $e');
-      setState(() {
-        showLoader = false;
-      });
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text(lang.translate('Error sending message')),
-          backgroundColor: Colors.red,
-        ),
-      );
-    }
+  void _navigateToCreateMessage() async {
+    // Navegar a la página de crear mensaje
+    await Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const CreateSupportMessagePage(),
+      ),
+    );
+    // Recargar mensajes al volver
+    await loadMessages();
+    setState(() {});
   }
 
   @override
@@ -183,71 +102,70 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
       return Loader();
     }
 
-    // Si no hay mensajes y no está creando, mostrar el formulario
-    if (messagesList!.isEmpty && !isCreatingMessage) {
-      return _buildFormView();
-    }
-
-    // Si está creando mensaje, mostrar el formulario
-    if (isCreatingMessage) {
-      return _buildFormView();
+    // Si no hay mensajes, mostrar estado vacío con botón para crear
+    if (messagesList!.isEmpty) {
+      return _buildEmptyState();
     }
 
     // Mostrar lista de mensajes
     return _buildMessagesListView();
   }
 
-  Widget _buildFormView() {
+  Widget _buildEmptyState() {
     return Scaffold(
       backgroundColor: Theme.of(context).scaffoldBackgroundColor,
       body: Stack(
         children: [
-          SingleChildScrollView(
-            padding: const EdgeInsets.only(top: 100, bottom: 20),
+          Center(
             child: Padding(
-              padding: const EdgeInsets.all(20.0),
+              padding: const EdgeInsets.all(40.0),
               child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  // Botón de volver si hay mensajes
-                  if (messagesList!.isNotEmpty) ...[
-                    IconButton(
-                      icon: Icon(Icons.arrow_back, color: activeTheme.main_color),
-                      onPressed: _toggleCreateMessage,
-                    ),
-                    const SizedBox(height: 20),
-                  ],
-
-                  // Título con animación
-                  AnimatedContainer(
-                    duration: const Duration(milliseconds: 500),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          lang.translate("We can help") + " 👋",
-                          style: TextStyle(
-                            fontSize: 32,
-                            fontWeight: FontWeight.bold,
-                            color: activeTheme.main_color,
-                          ),
-                        ),
-                        const SizedBox(height: 8),
-                        Text(
-                          lang.translate('send_your_message_below'),
-                          style: TextStyle(
-                            fontSize: 16,
-                            color: activeTheme.main_color.withOpacity(0.7),
-                          ),
-                        ),
-                      ],
+                  Icon(
+                    Icons.support_agent_rounded,
+                    size: 100,
+                    color: activeTheme.main_color.withOpacity(0.3),
+                  ),
+                  const SizedBox(height: 20),
+                  Text(
+                    lang.translate('no_messages'),
+                    style: TextStyle(
+                      fontSize: 24,
+                      fontWeight: FontWeight.bold,
+                      color: activeTheme.main_color,
                     ),
                   ),
-
+                  const SizedBox(height: 12),
+                  Text(
+                    lang.translate('send_your_message_below'),
+                    style: TextStyle(
+                      fontSize: 16,
+                      color: activeTheme.main_color.withOpacity(0.6),
+                    ),
+                    textAlign: TextAlign.center,
+                  ),
                   const SizedBox(height: 40),
-
-                  // Formulario con diseño mejorado
-                  _buildImprovedForm(),
+                  ElevatedButton.icon(
+                    onPressed: _navigateToCreateMessage,
+                    icon: Icon(Icons.add_comment_rounded, color: activeTheme.buttonColor),
+                    label: Text(
+                      lang.translate('new_message'),
+                      style: TextStyle(
+                        color: activeTheme.buttonColor,
+                        fontSize: 16,
+                        fontWeight: FontWeight.w600,
+                      ),
+                    ),
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor: activeTheme.buttonBG,
+                      padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 16),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                      elevation: 2,
+                    ),
+                  ),
                 ],
               ),
             ),
@@ -256,232 +174,6 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
         ],
       ),
     );
-  }
-
-  Widget _buildImprovedForm() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        // Categoría
-        _buildFormField(
-          label: lang.translate("Subject"),
-          child: Container(
-            decoration: BoxDecoration(
-              color: Theme.of(context).cardColor,
-              borderRadius: BorderRadius.circular(12),
-              border: Border.all(
-                color: activeTheme.main_color.withOpacity(0.2),
-                width: 1,
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.05),
-                  blurRadius: 10,
-                  offset: const Offset(0, 2),
-                ),
-              ],
-            ),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 12),
-              child: DropdownButtonHideUnderline(
-                child: DropdownButton<int>(
-                  value: categoryId,
-                  isExpanded: true,
-                  icon: Icon(Icons.expand_more, color: activeTheme.main_color),
-                  style: TextStyle(
-                    color: activeTheme.main_color,
-                    fontSize: 16,
-                  ),
-                  onChanged: (int? value) {
-                    setState(() {
-                      categoryId = value!;
-                    });
-                  },
-                  items: categories?.map<DropdownMenuItem<int>>(
-                    (SupportHelpCategory item) {
-                      return DropdownMenuItem<int>(
-                        value: item.id,
-                        child: Text(item.name!),
-                      );
-                    }
-                  ).toList() ?? [],
-                ),
-              ),
-            ),
-          ),
-        ),
-
-        const SizedBox(height: 24),
-
-        // Mensaje
-        _buildFormField(
-          label: lang.translate("Message"),
-          child: Container(
-            decoration: BoxDecoration(
-              color: Theme.of(context).cardColor,
-              borderRadius: BorderRadius.circular(12),
-              border: Border.all(
-                color: activeTheme.main_color.withOpacity(0.2),
-                width: 1,
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.05),
-                  blurRadius: 10,
-                  offset: const Offset(0, 2),
-                ),
-              ],
-            ),
-            child: TextField(
-              controller: _messageController,
-              maxLines: 5,
-              style: TextStyle(
-                color: activeTheme.main_color,
-                fontSize: 16,
-              ),
-              decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: const EdgeInsets.all(16),
-                hintText: lang.translate('Your message here'),
-                hintStyle: TextStyle(
-                  color: activeTheme.main_color.withOpacity(0.4),
-                ),
-              ),
-              onChanged: (val) => setState(() {
-                message = val;
-              }),
-            ),
-          ),
-        ),
-
-        const SizedBox(height: 24),
-
-        // Prioridad
-        _buildFormField(
-          label: lang.translate("Priority"),
-          child: Container(
-            decoration: BoxDecoration(
-              color: Theme.of(context).cardColor,
-              borderRadius: BorderRadius.circular(12),
-              border: Border.all(
-                color: activeTheme.main_color.withOpacity(0.2),
-                width: 1,
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.05),
-                  blurRadius: 10,
-                  offset: const Offset(0, 2),
-                ),
-              ],
-            ),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 12),
-              child: DropdownButtonHideUnderline(
-                child: DropdownButton<String>(
-                  value: priority,
-                  isExpanded: true,
-                  icon: Icon(Icons.expand_more, color: activeTheme.main_color),
-                  style: TextStyle(
-                    color: activeTheme.main_color,
-                    fontSize: 16,
-                  ),
-                  onChanged: (String? value) {
-                    setState(() {
-                      priority = value!;
-                    });
-                  },
-                  items: priorities.map<DropdownMenuItem<String>>((String value) {
-                    return DropdownMenuItem<String>(
-                      value: value,
-                      child: Row(
-                        children: [
-                          _getPriorityIcon(value),
-                          const SizedBox(width: 8),
-                          Text(lang.translate(value)),
-                        ],
-                      ),
-                    );
-                  }).toList(),
-                ),
-              ),
-            ),
-          ),
-        ),
-
-        const SizedBox(height: 40),
-
-        // Botón de enviar mejorado
-        Container(
-          width: double.infinity,
-          height: 56,
-          child: ElevatedButton(
-            onPressed: sendMessage,
-            style: ElevatedButton.styleFrom(
-              backgroundColor: activeTheme.buttonBG,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-              ),
-              elevation: 2,
-            ),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Icon(Icons.send_rounded, color: activeTheme.buttonColor),
-                const SizedBox(width: 8),
-                Text(
-                  lang.translate("Send now"),
-                  style: TextStyle(
-                    color: activeTheme.buttonColor,
-                    fontSize: 18,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildFormField({required String label, required Widget child}) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          label,
-          style: TextStyle(
-            fontSize: 14,
-            fontWeight: FontWeight.w600,
-            color: activeTheme.main_color,
-          ),
-        ),
-        const SizedBox(height: 8),
-        child,
-      ],
-    );
-  }
-
-  Widget _getPriorityIcon(String priority) {
-    IconData icon;
-    Color color;
-
-    switch (priority.toLowerCase()) {
-      case 'high':
-        icon = Icons.priority_high_rounded;
-        color = Colors.red;
-        break;
-      case 'low':
-        icon = Icons.low_priority_rounded;
-        color = Colors.blue;
-        break;
-      default:
-        icon = Icons.remove_rounded;
-        color = Colors.orange;
-    }
-
-    return Icon(icon, size: 20, color: color);
   }
 
   Widget _buildMessagesListView() {
@@ -579,12 +271,12 @@ class _SupportMessagesUnifiedPageState extends State<SupportMessagesUnifiedPage>
       floatingActionButton: ScaleTransition(
         scale: _fabAnimation,
         child: FloatingActionButton.extended(
-          onPressed: _toggleCreateMessage,
+          onPressed: _navigateToCreateMessage,
           backgroundColor: activeTheme.buttonBG,
           elevation: 4,
           icon: Icon(Icons.add_comment_rounded, color: activeTheme.buttonColor),
           label: Text(
-            lang.translate('New Message'),
+            lang.translate('new_message'),
             style: TextStyle(
               color: activeTheme.buttonColor,
               fontWeight: FontWeight.w600,
